treat FD_SIZE as int (which can be defined as unsigned on some platforms) to silence compiler. (checked by Mark)

+++ lib/lwres/context.c
@@ -15,7 +15,7 @@
  * PERFORMANCE OF THIS SOFTWARE.
  */
 
-/* $Id: context.c,v 1.50.332.3 2009/09/01 06:03:11 each Exp $ */
+/* $Id: context.c,v 1.50.332.4 2009/09/01 22:55:05 jinmei Exp $ */
 
 /*! \file context.c
    lwres_context_create() creates a #lwres_context_t structure for use in
@@ -476,7 +476,7 @@ lwres_context_sendrecv(lwres_context_t *ctx,
 	 * If this is not checked, select() can overflow,
 	 * causing corruption elsewhere.
 	 */
-	if (ctx->sock >= FD_SETSIZE) {
+	if (ctx->sock >= (int)FD_SETSIZE) {
 		close(ctx->sock);
 		ctx->sock = -1;
 		return (LWRES_R_IOERROR);
